feat: 최근 조회 페이지 클라이언트 컴포넌트 분리 및 api 응답 데이터 props 전달
